# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: AVN Hub API
  version: 0.2.5
  description: |
    REST API for the AVN Hub library server (`crates/api`).

    **Auth:** When an app password is configured, send `Authorization: Bearer <token>`
    (from `POST /api/v1/auth/login`). Unconfigured hubs accept unauthenticated calls.

    **Media:** Cached covers/screenshots are served under `/api/v1/media/{path}`.
    Clients should prefer `cached_url` on screenshots over F95 `full_url`.

    Keep this file in sync when adding or changing routes in `crates/api/src/routes.rs`.
servers:
  - url: http://127.0.0.1:8080
    description: Local / Docker API default
...
paths:
...
  /api/v1/auth/login:
    post:
      tags: [auth]
      summary: Exchange app password for session token
      operationId: authLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password:
                  type: string
      responses:
        "200":
          description: Session token
          content:
            application/json:
              schema:
                type: object
                required: [token]
                properties:
                  token:
                    type: string
  /api/v1/auth/logout:
    post:
      tags: [auth]
      summary: Invalidate current token
      operationId: authLogout
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkTrue"
  /api/v1/auth/me:
    get:
      tags: [auth]
      summary: Auth configuration and session state
      operationId: authMe
      security:
        - bearerAuth: []
        - {}
      responses:
        "200":
          description: Auth status
          content:
            application/json:
              schema:
                type: object
                required: [configured, authenticated]
                properties:
                  configured:
                    type: boolean
                  authenticated:
                    type: boolean
  /api/v1/settings:
    get:
      tags: [settings]
      summary: Get settings view
      operationId: getSettings
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettingsView"
    put:
      tags: [settings]
      summary: Update settings
      operationId: updateSettings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSettingsBody"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettingsView"
  /api/v1/settings/f95/login:
    post:
      tags: [settings]
      summary: Login to F95Zone with username/password (stores cookies)
      operationId: f95Login
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Result message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkMessage"
  /api/v1/settings/f95/cookies:
    get:
      tags: [settings]
      summary: Export stored F95 cookie header (desktop / Afterglow Browser)
      operationId: f95CookiesExport
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Cookie export
          content:
            application/json:
              schema:
                type: object
                required: [cookies, set]
                properties:
                  cookies:
                    type: string
                  set:
                    type: boolean
    post:
      tags: [settings]
      summary: Set F95 session cookies manually
      operationId: f95CookiesSet
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cookies]
              properties:
                cookies:
                  type: string
      responses:
        "200":
          description: Result message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkMessage"
  /api/v1/settings/storage:
    get:
      tags: [settings]
      summary: Storage usage stats
      operationId: storageStats
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StorageStats"
  /api/v1/settings/media/purge:
    post:
      tags: [settings]
      summary: Purge on-disk media cache
      operationId: purgeMedia
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Purged
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkTrue"
  /api/v1/catalog/search:
    get:
      tags: [catalog]
      summary: Search F95Zone catalog
      operationId: catalogSearch
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/catalogQ"
        - $ref: "#/components/parameters/catalogSearch"
        - $ref: "#/components/parameters/catalogCreator"
        - $ref: "#/components/parameters/catalogPage"
        - $ref: "#/components/parameters/catalogRows"
        - $ref: "#/components/parameters/catalogSort"
        - $ref: "#/components/parameters/catalogDate"
        - $ref: "#/components/parameters/catalogTagMode"
        - $ref: "#/components/parameters/catalogTags"
        - $ref: "#/components/parameters/catalogNotags"
        - $ref: "#/components/parameters/catalogPrefixes"
      responses:
        "200":
          description: Search page with SAM pagination totals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogPage"
  /api/v1/catalog/browse:
    get:
      tags: [catalog]
      summary: Browse F95Zone catalog (same filter shape as search)
      operationId: catalogBrowse
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/catalogQ"
        - $ref: "#/components/parameters/catalogSearch"
        - $ref: "#/components/parameters/catalogCreator"
        - $ref: "#/components/parameters/catalogPage"
        - $ref: "#/components/parameters/catalogRows"
        - $ref: "#/components/parameters/catalogSort"
        - $ref: "#/components/parameters/catalogDate"
        - $ref: "#/components/parameters/catalogTagMode"
        - $ref: "#/components/parameters/catalogTags"
        - $ref: "#/components/parameters/catalogNotags"
        - $ref: "#/components/parameters/catalogPrefixes"
      responses:
        "200":
          description: Browse page with SAM pagination totals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogPage"
  /api/v1/catalog/preview:
    get:
      tags: [catalog]
      summary: Preview a thread before adding (description + gallery)
      operationId: catalogPreview
      security:
        - bearerAuth: []
      parameters:
        - name: input
          in: query
          required: true
          schema:
            type: string
          description: F95 thread URL or numeric id
      responses:
        "200":
          description: Preview metadata with overview and screenshots
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CatalogPreview"
  /api/v1/catalog/tags:
    get:
      tags: [catalog]
      summary: Tag catalog autocomplete
      operationId: catalogTags
      security:
        - bearerAuth: []
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 200
      responses:
        "200":
          description: Tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CatalogTag"
  /api/v1/library:
    get:
      tags: [library]
      summary: List library games
      operationId: listLibrary
      security:
        - bearerAuth: []
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: play_status
          in: query
          schema:
            type: string
            enum: [unplayed, playing, completed, dropped]
        - name: user_rating
          in: query
          schema:
            type: string
          description: Minimum rating (number as string) or `unrated`
        - name: user_rating_min
          in: query
          schema:
            type: number
        - name: tags
          in: query
          schema:
            type: string
          description: Comma-separated tags
        - name: tag_mode
          in: query
          schema:
            type: string
            enum: [and, or]
            default: and
        - name: platforms
          in: query
          schema:
            type: string
          description: Comma-separated platforms
        - name: platform_mode
          in: query
          schema:
            type: string
            enum: [and, or]
            default: or
        - name: sort
          in: query
          schema:
            type: string
            enum:
              - title_asc
              - title_desc
              - updated_desc
              - rating_desc
              - user_rating_desc
      responses:
        "200":
          description: Library
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GameSummary"
  /api/v1/library/tags:
    get:
      tags: [library]
      summary: Distinct tags in library with counts
      operationId: libraryTags
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LibraryTag"
  /api/v1/library/platforms:
    get:
      tags: [library]
      summary: Distinct platforms in library with counts
      operationId: libraryPlatforms
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Platforms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LibraryPlatform"
  /api/v1/library/add:
    post:
      tags: [library]
      summary: Add game from F95 thread URL or id
      operationId: addGame
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [input]
              properties:
                input:
                  type: string
      responses:
        "200":
          description: Added game detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameDetail"
  /api/v1/library/check-updates:
    post:
      tags: [library]
      summary: Check all games for F95 version updates
      operationId: checkAllUpdates
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Per-game version results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/VersionCheckResult"
  /api/v1/games/{id}:
    parameters:
      - $ref: "#/components/parameters/gameId"
    get:
      tags: [games]
      summary: Game detail (metadata, screenshots, saves, patches)
      operationId: getGame
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameDetail"
    patch:
      tags: [games]
      summary: Update user tracking fields
      description: |
        Partial update. Omitted fields are left unchanged.

        For `user_rating`:
        - omit field → leave unchanged
        - `null` → clear rating
        - number (0.5–5, half steps) → set rating
      operationId: patchGame
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateGameUserData"
      responses:
        "200":
          description: Updated detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameDetail"
    delete:
      tags: [games]
      summary: Remove game from library
      operationId: deleteGame
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: session token
  parameters:
...
    catalogQ:
      name: q
      in: query
      schema:
        type: string
    catalogSearch:
      name: search
      in: query
      schema:
        type: string
    catalogCreator:
      name: creator
      in: query
      schema:
        type: string
...
  schemas:
    OkTrue:
      type: object
      properties:
        ok:
          type: boolean
          const: true
    OkMessage:
      type: object
      properties:
        ok:
          type: boolean
        message:
          type: string
...
    SettingsView:
      type: object
      properties:
        data_dir:
          type: string
        f95_username:
          type: [string, "null"]
        f95_password_set:
          type: boolean
        f95_cookies_set:
          type: boolean
        f95_authenticated:
          type: boolean
        app_password_set:
          type: boolean
        max_attachment_bytes:
          type: integer
          format: int64
        tag_click_action:
          type: string
        save_sync_enabled:
          type: boolean
        save_sync_max_per_game:
          type: integer
        save_sync_rolling:
          type: boolean
        save_sync_name_format:
          type: string
...